refact: add new FC ButtonsFilter
